+++ tests/test_setup/test_nu.py
